+++ src/components/navigation/document-tab-content.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { DocumentModelType } from "../../models/document/document";
-import { NavTabSpec } from "../../models/view/left-tabs";
+import { NavTabSpec } from "../../models/view/nav-tabs";
 import { DocumentTabPanel } from "./document-tab-panel";
 import { EditableDocumentContent } from "../document/editable-document-content";
