+++ src/components/navigation/document-tab-panel.tsx
@@ -2,7 +2,7 @@ import { inject, observer } from "mobx-react";
 import React from "react";
 import { BaseComponent, IBaseProps } from "../base";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
-import { ENavTabSectionType, NavTabSpec, NavTabSectionModelType  } from "../../models/view/left-tabs";
+import { ENavTabSectionType, NavTabSpec, NavTabSectionModelType  } from "../../models/view/nav-tabs";
 import { IStores } from "../../models/stores/stores";
 import { TabPanelDocumentsSection } from "../thumbnail/tab-panel-documents-section";
 import { DocumentDragKey, DocumentModelType, SupportPublication } from "../../models/document/document";
