+++ src/models/view/nav-tabs.ts
